+++ blocklets/ai-studio/src/components/file-editor/input/InputTable.tsx
@@ -624,15 +624,15 @@ function SelectFromSourceDialog({
 
             if (parameter.type === 'source' && parameter?.source?.variableFrom === 'knowledge' && parameter?.source) {
               const { source } = parameter;
-              Object.entries(source?.knowledge?.parameters || {}).forEach(([key, value]: any) => {
-                if (!value) {
-                  if (source && source?.knowledge && source?.knowledge?.parameters) {
-                    source.knowledge.parameters[key] = `{{${key}}}`;
-                  }
-
-                  addParameter(key, { from: FROM_KNOWLEDGE_PARAMETER });
-                }
-              });
+              if (
+                source &&
+                source?.knowledge &&
+                source?.knowledge?.parameters &&
+                !source?.knowledge?.parameters?.message
+              ) {
+                source.knowledge.parameters.message = '{{message}}';
+                addParameter('message', { from: FROM_KNOWLEDGE_PARAMETER });
+              }
             }
 
             dialogState.close();
@@ -751,7 +751,7 @@ function KnowledgeParameter({
     ];
 
     const parameters = [
-      { name: 'all', description: t('allContent'), type: 'boolean' },
+      { name: 'searchAll', description: t('allContent'), type: 'boolean' },
       { name: 'message', description: t('searchContent') },
     ];
     const v = options.find((x) => x.id === toolId);
@@ -770,22 +770,23 @@ function KnowledgeParameter({
               if (_value) {
                 // 删除历史自动添加的变量
                 Object.values(value.parameters || {}).forEach((x) => {
-                  if (x.data.from === FROM_KNOWLEDGE_PARAMETER) {
+                  if (
+                    (x.data.from === FROM_KNOWLEDGE_PARAMETER || x.data.from === FROM_PARAMETER) &&
+                    !checkKeyParameterIsUsed({ value, key: x.data.key || '' })
+                  ) {
                     deleteParameter(x.data);
                   }
                 });
 
-                if (!source?.knowledge?.parameters?.all) {
-                  const parameters = {
-                    message: '',
-                  };
+                const parameters = {
+                  message: '',
+                };
 
-                  source.knowledge = {
-                    id: _value.id,
-                    from: 'knowledge',
-                    parameters,
-                  };
-                }
+                source.knowledge = {
+                  id: _value.id,
+                  from: 'knowledge',
+                  parameters,
+                };
               }
             }}
           />
@@ -817,7 +818,7 @@ function KnowledgeParameter({
                   );
                 }
 
-                if (source?.knowledge?.parameters?.all) {
+                if (source?.knowledge?.parameters?.searchAll) {
                   return null;
                 }
 
@@ -896,6 +897,33 @@ function DatastoreParameter({
 
 const filter = createFilterOptions<any>();
 
+function checkKeyParameterIsUsed({ value, key }: { value: AssistantYjs; key: string }) {
+  if (!key) {
+    return false;
+  }
+
+  const parameters = Object.values(value?.parameters || {}).flatMap((x) => {
+    if (x.data.type === 'source') {
+      if (x.data.source?.variableFrom === 'tool') {
+        return [Object.values(x.data.source?.agent?.parameters || {})];
+      }
+
+      if (x.data.source?.variableFrom === 'knowledge') {
+        return [Object.values(x.data.source?.knowledge?.parameters || {})];
+      }
+    }
+
+    return [];
+  });
+
+  return !!parameters.find((x) => {
+    return (x || []).find((str) => {
+      const pattern = new RegExp(`{{\\s*${key}\\s*}}`);
+      return pattern.test(str);
+    });
+  });
+}
+
 function AgentParameter({
   value,
   projectId,
@@ -944,7 +972,10 @@ function AgentParameter({
               if (_value) {
                 // 删除历史自动添加的变量
                 Object.values(value.parameters || {}).forEach((x) => {
-                  if (x.data.from === FROM_PARAMETER) {
+                  if (
+                    (x.data.from === FROM_KNOWLEDGE_PARAMETER || x.data.from === FROM_PARAMETER) &&
+                    !checkKeyParameterIsUsed({ value, key: x.data.key || '' })
+                  ) {
                     deleteParameter(x.data);
                   }
                 });