+++ blocklets/ai-studio/src/components/file-editor/output/OutputSettings.tsx
@@ -266,7 +266,14 @@ function VariableRow({
     return `${x.scope}_${x.key}` === `${j.scope}_${j.key}`;
   });
 
-  const v = datastoreVariable?.type
+  const error = useCheckConflictAssistantOutputAndSelectAgents({
+    selectAgentOutputVariables,
+    value,
+    v: variable,
+    depth,
+  });
+
+  const mergeVariable = datastoreVariable?.type
     ? {
         ...datastoreVariable?.type,
         id: variable.id,
@@ -275,8 +282,6 @@ function VariableRow({
         required: variable.required,
       }
     : variable;
-
-  const error = useCheckConflictAssistantOutputAndSelectAgents({ selectAgentOutputVariables, value, v, depth });
 
   return (
     <>
@@ -289,11 +294,14 @@ function VariableRow({
           }}>
           <Box component={TableCell}>
             <Box sx={{ ml: depth }}>
-              <OutputNameCell output={v} TextFieldProps={{ disabled: Boolean(disabled) || parent?.type === 'array' }} />
+              <OutputNameCell
+                output={variable}
+                TextFieldProps={{ disabled: Boolean(disabled) || parent?.type === 'array' }}
+              />
             </Box>
           </Box>
           <Box component={TableCell}>
-            <OutputDescriptionCell assistant={value} output={v} TextFieldProps={{ disabled }} />
+            <OutputDescriptionCell assistant={value} output={variable} TextFieldProps={{ disabled }} />
           </Box>
           <Box component={TableCell}>
             <OutputFormatCell output={variable} variable={datastoreVariable} TextFieldProps={{ disabled }} />
@@ -303,9 +311,9 @@ function VariableRow({
               <Switch
                 size="small"
                 disabled={Boolean(disabled)}
-                checked={v.required || false}
+                checked={variable.required || false}
                 onChange={(_, checked) => {
-                  v.required = checked;
+                  variable.required = checked;
                 }}
               />
             )}
@@ -326,12 +334,12 @@ function VariableRow({
       </Tooltip>
 
       {!runtimeVariable &&
-        v.type === 'object' &&
-        v.properties &&
-        sortBy(Object.values(v.properties), 'index').map((property) => (
+        mergeVariable.type === 'object' &&
+        mergeVariable.properties &&
+        sortBy(Object.values(mergeVariable.properties), 'index').map((property) => (
           <React.Fragment key={property.data.id}>
             <VariableRow
-              parent={v}
+              parent={mergeVariable}
               disabled={Boolean(variable.variable?.key || disabled)}
               value={value}
               variable={property.data}
@@ -341,23 +349,25 @@ function VariableRow({
               onRemove={() => {
                 const doc = (getYjsValue(variable) as Map<any>).doc!;
                 doc.transact(() => {
-                  if (!v.properties) return;
-                  delete v.properties[property.data.id];
-                  sortBy(Object.values(v.properties), 'index').forEach((item, index) => (item.index = index));
+                  if (!mergeVariable.properties) return;
+                  delete mergeVariable.properties[property.data.id];
+                  sortBy(Object.values(mergeVariable.properties), 'index').forEach(
+                    (item, index) => (item.index = index)
+                  );
                 });
               }}
             />
           </React.Fragment>
         ))}
 
-      {!runtimeVariable && v.type === 'array' && v.element && (
+      {!runtimeVariable && mergeVariable.type === 'array' && mergeVariable.element && (
         <VariableRow
-          parent={v}
+          parent={mergeVariable}
           disabled={Boolean(variable.variable?.key || disabled)}
           projectId={projectId}
           gitRef={gitRef}
           value={value}
-          variable={v.element}
+          variable={mergeVariable.element}
           depth={depth + 1}
         />
       )}